+++ inbm/Changelog.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## NEXT - ?
 
+### Added
+ - Added INBS cloud with Ping support to Cloudadapter
+
 ### Changed
  - Moved service files from /lib to /usr/lib for all Linux OSes